relax parsing of size validation
rather than pedantically check for lengths to be in minimum form, parse
them if they are parseable.

NetApp's encoding with SPNEGO dumbly encodes _all_ size values as 02 HH
LL regardless of size. This may be in violation of ASN.1, but it doesn't
break Windows so doubtful NetApp would be in any hurry to fix it.

I've left the base128 side of things alone.